conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
anyio=4.2.0=py312hca03da5_0
appnope=0.1.3=py312hca03da5_1001
arrow=1.2.3=py312hca03da5_1
asttokens=2.0.5=pyhd3eb1b0_0
attrs=23.1.0=py312hca03da5_0
binaryornot=0.4.4=pyhd3eb1b0_1
blas=1.0=openblas
bottleneck=1.3.7=py312ha86b861_0
brotli=1.0.9=h80987f9_8
brotli-bin=1.0.9=h80987f9_8
brotli-python=1.0.9=py312h313beb8_8
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.3.11=hca03da5_0
cachetools=5.3.3=py312hca03da5_0
certifi=2024.2.2=py312hca03da5_0
chardet=4.0.0=py312hca03da5_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py312hca03da5_0
comm=0.2.1=py312hca03da5_0
contourpy=1.2.0=py312h48ca7d4_0
cookiecutter=2.6.0=py312hca03da5_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py312h313beb8_0
decorator=5.1.1=pyhd3eb1b0_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.6.2=h313beb8_0
filelock=3.13.1=py312hca03da5_0
fonttools=4.51.0=py312h80987f9_0
freetype=2.12.1=h1192e45_0
h11=0.14.0=py312hca03da5_0
humanize=3.10.0=pyhd3eb1b0_0
idna=3.7=py312hca03da5_0
ipykernel=6.28.0=py312hca03da5_0
ipython=8.20.0=py312hca03da5_0
ipyvue=1.11.0=pyhd8ed1ab_0
ipyvuetify=1.9.4=pyhd8ed1ab_0
ipywidgets=8.1.2=py312hca03da5_0
jedi=0.18.1=py312hca03da5_1
jinja2=3.1.3=py312hca03da5_0
jpeg=9e=h80987f9_1
jsonschema=4.19.2=py312hca03da5_0
jsonschema-specifications=2023.7.1=py312hca03da5_0
jupyter_client=8.6.0=py312hca03da5_0
jupyter_core=5.5.0=py312hca03da5_0
jupyterlab_widgets=3.0.10=py312hca03da5_0
kiwisolver=1.4.4=py312h313beb8_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libbrotlicommon=1.0.9=h80987f9_8
libbrotlidec=1.0.9=h80987f9_8
libbrotlienc=1.0.9=h80987f9_8
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_1
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libsodium=1.0.18=h1a28f6b_0
libtiff=4.5.1=h313beb8_0
libwebp-base=1.3.2=h80987f9_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_1
markdown=3.4.1=py312hca03da5_0
markdown-it-py=2.2.0=py312hca03da5_1
markupsafe=2.1.3=py312h80987f9_0
matplotlib-base=3.8.4=py312hd77ebd4_0
matplotlib-inline=0.1.6=py312hca03da5_0
mdurl=0.1.0=py312hca03da5_0
mesa=2.3.0=pyhd8ed1ab_0
mesa-viz-tornado=0.1.3=pyhd8ed1ab_0
nbformat=5.9.2=py312hca03da5_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=py312hca03da5_0
networkx=3.1=py312hca03da5_0
numexpr=2.8.7=py312h0f3ea24_0
numpy=1.26.4=py312h7f4fdc5_0
numpy-base=1.26.4=py312he047099_0
openjpeg=2.3.0=h7a6adac_2
openssl=3.0.13=h1a28f6b_1
packaging=23.2=py312hca03da5_0
pandas=2.2.1=py312hd77ebd4_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pillow=10.3.0=py312h80987f9_0
pip=24.0=py312hca03da5_0
platformdirs=3.10.0=py312hca03da5_0
prompt-toolkit=3.0.43=py312hca03da5_0
prompt_toolkit=3.0.43=hd3eb1b0_0
psutil=5.9.0=py312h80987f9_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pygments=2.15.1=py312hca03da5_1
pymdown-extensions=9.11=py312hca03da5_0
pyparsing=3.0.9=py312hca03da5_0
pysocks=1.7.1=py312hca03da5_0
python=3.12.3=h99e199e_1
python-dateutil=2.9.0post0=py312hca03da5_0
python-fastjsonschema=2.16.2=py312hca03da5_0
python-slugify=5.0.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=py312hca03da5_0
pyyaml=6.0.1=py312h80987f9_0
pyzmq=25.1.2=py312h313beb8_0
reacton=1.8.3=pyhd8ed1ab_0
readline=8.2=h1a28f6b_0
referencing=0.30.2=py312hca03da5_0
requests=2.31.0=py312hca03da5_1
rich=13.3.5=py312hca03da5_1
rich-click=1.8.2=pyhd8ed1ab_0
rpds-py=0.10.6=py312hf0e4da2_0
setuptools=69.5.1=py312hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=py312hca03da5_0
solara=1.23.0=pyhd8ed1ab_0
sqlite=3.45.3=h80987f9_0
stack_data=0.2.0=pyhd3eb1b0_0
starlette=0.27.0=py312hca03da5_0
text-unidecode=1.3=pyhd3eb1b0_0
tk=8.6.14=h6ba3021_0
tornado=6.3.3=py312h80987f9_0
tqdm=4.66.2=py312h989b03a_0
traitlets=5.7.1=py312hca03da5_0
typing_extensions=4.11.0=py312hca03da5_0
tzdata=2024a=h04d1e81_0
unicodedata2=15.1.0=py312h80987f9_0
unidecode=1.2.0=pyhd3eb1b0_0
urllib3=2.2.1=py312hca03da5_0
uvicorn=0.20.0=py312hca03da5_0
watchdog=2.1.6=py312h80987f9_0
wcwidth=0.2.5=pyhd3eb1b0_0
websockets=10.4=py312h80987f9_1
wheel=0.43.0=py312hca03da5_0
widgetsnbextension=4.0.10=py312hca03da5_0
xz=5.4.6=h80987f9_1
yaml=0.2.5=h1a28f6b_0
zeromq=4.3.5=h313beb8_0
zlib=1.2.13=h18a0788_1
zstd=1.5.5=hd90d995_2
